Fix F24.csv
I'm not sure if this is the right way to fix this. Where should the missing commas be inserted?

Row 12 I feel good about, because
that's  pretty normal for the old versions
to just have a _name field,
and then in newer versions they
split it into _last_name, _first_name,
etc.

Row 5, I'm not as sure. Perhaps the commas should be earlier.